packages/gemma-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grantex-gemma"
version = "0.1.0"
description = "Offline authorization for Gemma 4 on-device AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{ name = "Sanjeev Kumar" }]
keywords = [
    "grantex",
    "gemma",
    "offline-auth",
    "ai-agents",
    "jwt",
    "on-device",
    "authorization",
    "audit",
    "hash-chain",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "PyJWT>=2.8",
    "cryptography>=41",
]

[project.urls]
Homepage = "https://grantex.dev"
Documentation = "https://docs.grantex.dev"
Repository = "https://github.com/mishrasanjeev/grantex"
Issues = "https://github.com/mishrasanjeev/grantex/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "mypy>=1.9",
    "ruff>=0.4",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/grantex_gemma"]

[tool.mypy]
strict = true
files = ["src/grantex_gemma"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "live: tests that hit the live production Grantex service (requires network)",
]

[tool.ruff]
line-length = 88
target-version = "py39"
